cretaed a new package helpClasses and moved the help classes from factories to there
