+++ src/main/java/com/example/StayNest/helpClasses/ConvertToBookingResponseDTO.java
@@ -1,4 +1,4 @@
-package com.example.StayNest.factories;
+package com.example.StayNest.helpClasses;
 
 import com.example.StayNest.dto.BookingResponseDTO;
 import com.example.StayNest.models.Booking;
