+++ src/main/java/com/example/StayNest/services/BookingService.java
@@ -4,8 +4,8 @@ import com.example.StayNest.dto.BookingResponseDTO;
 import com.example.StayNest.exceptions.ResourceNotFoundException;
 import com.example.StayNest.exceptions.UnauthorizedException;
 import com.example.StayNest.factories.BookingFactory;
-import com.example.StayNest.factories.CalculateTotalAmount;
-import com.example.StayNest.factories.ConvertToBookingResponseDTO;
+import com.example.StayNest.helpClasses.CalculateTotalAmount;
+import com.example.StayNest.helpClasses.ConvertToBookingResponseDTO;
 import com.example.StayNest.models.Booking;
 import com.example.StayNest.models.User;
 import com.example.StayNest.repositories.BookingRepository;
